I've added some mock ws connection events.
I'm not sure how to go about using them. I tried to use them in the `invoke-req-test` but that test basically calls a function that does nothing for websocket events and returns everything verbatim back to the mock invoke. I'm testing the input is matching the output, which it will for the most part. I could see using a form of these mocks in an integration test (like test/integration/http/websockets-test.js). But I want some guidance.