feat: make a request context for websocket events (#598)
* feat: make a request context for websocket events

I hit a bug where requestContext.eventType wasn't populated and lost a few hours of debugging to it. It turns out that some websocket libraries require this.

I used logs from actual websocket events to design the requestContexts.

bonus:
- Made a little request-id helper that should give us some nice random request ids.

* lint: fix some consts for lets

* I've added some mock ws connection events.

I'm not sure how to go about using them. I tried to use them in the `invoke-req-test` but that test basically calls a function that does nothing for websocket events and returns everything verbatim back to the mock invoke. I'm testing the input is matching the output, which it will for the most part. I could see using a form of these mocks in an integration test (like test/integration/http/websockets-test.js). But I want some guidance.

* fix: wrong event type and route key on default messages

* fix: have a valid domain name for the sandbox

* Add stage

* rollback test changes

* fix: tests because the call signature of invoke has changed

* chore: complete rewrite of the websocket integration test

This test ensures all message type events have correct inputs when the function is executed

* chore: use an http sidechannel to catalog all websocket events

* chore: change sidechannel to use promises and not async generators

This is a better api

* Use the sidechannel for webscoket tests and remove duplicate code

* remove test logic from the sidechannel

It's super annoying that the tests end when the last test is called in a plan as opposed to when the async function resolves. Tape says async functions are cool but that is uncool.

* fix node 10, they handle the close vs finish events differently

* remove consts

* fix: pr feedback

* Pass all the info we need to make the request context around from the places we need it

* Dry up remaining @ws invocation request formatting

Co-authored-by: Ryan Block