fix: shutting down the sandbox still fires websocket disconnect lambda
Connections were emitting the close event which would invoke the disconnect lambda when the sandbox was shutdown. This might not be awful but at this point we've usually also shutdown the service discovery, the DDB, the primary http server and SNS. So it's not great and has led to a lot of strange looking errors and crashes, especially when used in tests where you often don't care that every client has been disconnected before shutdown.