Adds host flag for servers to listen on
It's useful to be able to specify the address for the services to listen
on, this is espescially useful in situations where you might have the
sandbox inside of a docker network during development and want other
containers (e.g. a proxy webserver) to be able to communicate with arc.